feat(table): moving files in new folders, change in structure
