+++ packages/mantine/src/components/table/table-column/TableCollapsibleColumn.tsx
