+++ packages/mantine/src/components/table/table-footer/TableFooter.tsx
