+++ packages/mantine/src/components/table/table-header/Th.tsx
