+++ packages/mantine/src/components/table/table-pagination/TablePagination.tsx
@@ -1,7 +1,7 @@
 import {Pagination} from '@mantine/core';
 import {FunctionComponent} from 'react';
 
-import {useTable} from './TableContext';
+import {useTable} from '../TableContext';
 
 interface TablePaginationProps {
     /**
